Another horrible partial commit
Going home. Very distracted by the ICCV deadline. Started
working on code to automatically find matlab and cache the
path on the config file. But that requires refactoring the
config file code so that it can support multiple options.
That's about as far as I got - config code needs a bunch
of work.